// projects/modules/node-snmp/src/lib/types/connection-snmp.ts
import {Connection} from '@anyopsos/backend-core/app/types/connection';
import {DataObject} from '@anyopsos/backend-core/app/types/data-object';

export interface ConnectionSnmp extends Connection {
  type: 'snmp';
  host: string;
  port: number;
  trapPort: number;
  version: 'v1' | 'v2c' | 'v3'
  community: string;
  hopServerUuid: string;
  data: {
    Base: {
      name: string;
    };
    Data: (DataObject & { info: { data: any } })[];
  };
}